removed comments

+++ app/api/users/[id]/route.ts
@@ -4,7 +4,6 @@ import connectToDatabase from "@/utils/database";
 import { getServerSession, Session } from "next-auth";
 import { GET as AuthGET } from "@/app/api/auth/[...nextauth]/route";
 
-//get
 export const GET = async (req: NextRequest, { params }: { params: { id: string } }) => {
   const id = params.id;
   try {
@@ -19,11 +18,11 @@ export const GET = async (req: NextRequest, { params }: { params: { id: string }
 
 export const DELETE = async (req: NextRequest, { params }: { params: { id: string } }) => {
   const id = params.id;
-  //const session: Session = await getServerSession(AuthGET)
+  const session: Session = await getServerSession(AuthGET);
   try {
-    // if (!session || session.user.email !== process.env.ADMIN_EMAIL) {
-    //   return NextResponse.json("Unauthorized", { status: 401 })
-    // }
+    if (!session || session.user.email !== process.env.ADMIN_EMAIL) {
+      return NextResponse.json("Unauthorized", { status: 401 });
+    }
     await connectToDatabase();
     const user = await User.findByIdAndDelete(id);
     return NextResponse.json(user, { status: 200 });
@@ -33,15 +32,14 @@ export const DELETE = async (req: NextRequest, { params }: { params: { id: strin
   }
 };
 
-// update example we wont have update for users in this app just an example
 export const PATCH = async (req: NextRequest, { params }: { params: { id: string } }) => {
   const id = params.id;
   const { email, username, image, role } = await req.json();
   const session: Session = await getServerSession(AuthGET);
   try {
-    // if (!session || session.user.email !== process.env.ADMIN_EMAIL) {
-    //   return NextResponse.json("Unauthorized", { status: 401 })
-    // }
+    if (!session || session.user.email !== process.env.ADMIN_EMAIL) {
+      return NextResponse.json("Unauthorized", { status: 401 });
+    }
     await connectToDatabase();
     const user = await User.findByIdAndUpdate(id, { email, username, image, role }, { new: true });
     return NextResponse.json(user, { status: 200 });
